# api-doc.json
{
  "swagger": "2.0",
  "info": {
    "title": "Survey Node API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Login",
        "description": "",
        "operationId": "adminLogin",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful."
          }
        }
      }
    },
    "/api/admin/me": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin User",
        "description": "",
        "operationId": "adminUser",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User is authenticated."
          }
        }
      }
    },
    "/api/admin/signup": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Signup",
        "description": "",
        "operationId": "adminSignup",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password",
                "username"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signup successful."
          }
        }
      }
    },
    "/api/admin/update": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Update",
        "description": "",
        "operationId": "adminUpdate",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password",
                "username"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update successful."
          }
        }
      }
    },
    "/api/input_options/{id}": {
      "get": {
        "tags": [
          "Input Options"
        ],
        "summary": "Get Input Options",
        "description": "",
        "operationId": "inputOptions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Input option"
          }
        }
      },
      "put": {
        "tags": [
          "Input Options"
        ],
        "summary": "Update Input Options",
        "description": "",
        "operationId": "updateInputOption",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "label"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Input option"
          }
        }
      },
      "delete": {
        "tags": [
          "Input Options"
        ],
        "summary": "Update Input Options",
        "description": "",
        "operationId": "updateInputOption",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Input option deleted"
          }
        }
      }
    },
    "/api/input_options/create": {
      "post": {
        "tags": [
          "Input Options"
        ],
        "summary": "Create Input Options",
        "description": "",
        "operationId": "createInputOption",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "label"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Input option"
          }
        }
      }
    },
    "/api/input_options/": {
      "get": {
        "tags": [
          "Input Options"
        ],
        "summary": "Get Input Options",
        "description": "",
        "operationId": "getInputOptions",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Input option"
          }
        }
      }
    },
    "/api/input_types/{id}": {
      "get": {
        "tags": [
          "Input Types"
        ],
        "summary": "Get Input Types",
        "description": "",
        "operationId": "inputTypes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Input type"
          }
        }
      },
      "put": {
        "tags": [
          "Input Types"
        ],
        "summary": "Update Input Types",
        "description": "",
        "operationId": "updateInputType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "label"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Input type"
          }
        }
      },
      "delete": {
        "tags": [
          "Input Types"
        ],
        "summary": "Update Input Types",
        "description": "",
        "operationId": "updateInputType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Input type deleted"
          }
        }
      }
    },
    "/api/input_types/create": {
      "post": {
        "tags": [
          "Input Types"
        ],
        "summary": "Create Input Types",
        "description": "",
        "operationId": "createInputType",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "label"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Input type"
          }
        }
      }
    },
    "/api/input_types/": {
      "get": {
        "tags": [
          "Input Types"
        ],
        "summary": "Get Input Types",
        "description": "",
        "operationId": "getInputTypes",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Input type"
          }
        }
      }
    },
...
  }
}